implements a UrlService
if you want to get a URL for RDS, you can use this class to interact with
It makes the RDS URLs maintainable and reduce writing afford.
Also it implements an app-wide config value, which can be configured from the admin.
The value can be set and get with this class.