layer0_web: Use same nameing scheme as helm chart
We found a bug in the helm chart[0] where the naming scheme of the
environent variable differed from the one used in the app here. This
patch will switch to use a similar, more comprehensive, method of
removing unwanted characters from the name of the environment variables.
This means that we should get the same result from the chart and the app
to the extent that golang and python agrees on regexes.

Now all characters that ar not in the set [a-zA-Z0-9_] should be
replaced with underscore (_).

Notes:
 0. https://github.com/Sciebo-RDS/charts/pull/9

Signed-off-by: Micke Nordin <kano@sunet.se>